backend/notamil-backend/pyproject.toml
[tool.poetry]
name = "notamil-backend"
version = "0.1.0"
description = "Redato Backend"
authors = ["Wenzo <rithellyenzo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
gunicorn = "^23.0.0"
sse-starlette = "2.2.1"
aiohttp = "^3.11.11"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
python-multipart = "^0.0.20"
pydantic = {extras = ["email"], version = "^2.10.6"}
firebase-admin = "^6.6.0"
google-cloud-logging = "^3.11.3"
google-cloud-bigquery = "^3.29.0"
google-cloud-secret-manager = "^2.22.1"
pytz = "^2024.2"
functions-framework = "^3.8.2"
requests = "^2.32.3"
openai = "^1.60.1"
spacy = "3.6.1"
google-cloud-vision = "^3.9.0"
anthropic = "^0.45.0"
numpy = "1.26.0"
isort = "5.13.2"
black = "22.12.0"
flake8 = "3.9.2"
opencv-python-headless = "4.8.0.76"
redis = "^5.2.1"
litellm = "^1.63.14"
tenacity = "^9.0.0"
openpyxl = "^3.1.5"
pandas = "^2.2.3"

[tool.poetry.group.dev.dependencies]
flake8 = "3.9.2"
black = "22.12.0"

[tool.isort]
profile = "black"
line_length = 90
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
combine_as_imports = true
lines_between_types = 1
src_paths = ["iae_backend"]

[tool.black]
line-length = 90
target-version = ['py39']
include = '\.pyi?$'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[build.env]
name = "GOOGLE_RUNTIME_VERSION"
value = "3.9"

